Change delete message if successful, but react with emotes if unsuccessful

+++ main.py
@@ -163,8 +163,12 @@ class MyClient(discord.Client):
             message.channel.send('Neplatný stav světla. Správně je: **!light [on|off]**')
             return True
         light_state = 1 if message.content.lower() == '!light on' else 0
-        socket_client.switch_light(SOCKET_SERVER_ADDRESS, light_state)
-        await message.delete()
+        sent = await socket_client.switch_light(SOCKET_SERVER_ADDRESS, light_state)
+        if sent: await message.delete()
+        else:
+            await message.add_reaction(EMOTES["confused"])
+            await message.add_reaction(EMOTES["thumbsdown"])
+        
         return True
